Converted deployment targets into availability attributes.

+++ Sources/CCryptoBoringSSL/crypto/asn1/tasn_dec.c
@@ -456,7 +456,7 @@ static int asn1_item_ex_d2i(ASN1_VALUE **pval, const unsigned char **in,
             }
         }
         /* Save encoding */
-        if (!asn1_enc_save(pval, *in, p - *in, it))
+        if (!asn1_enc_save(pval, *in, (int)(p - *in), it))
             goto auxerr;
         if (asn1_cb && !asn1_cb(ASN1_OP_D2I_POST, pval, it, NULL))
             goto auxerr;
@@ -503,7 +503,7 @@ static int asn1_template_ex_d2i(ASN1_VALUE **val,
     char exp_eoc;
     if (!val)
         return 0;
-    flags = tt->flags;
+    flags = (int)tt->flags;
     aclass = flags & ASN1_TFLG_TAG_CLASS;
 
     p = *in;
@@ -516,7 +516,7 @@ static int asn1_template_ex_d2i(ASN1_VALUE **val,
          * where it starts: so read in EXPLICIT header to get the info.
          */
         ret = asn1_check_tlen(&len, NULL, NULL, &exp_eoc, &cst,
-                              &p, inlen, tt->tag, aclass, opt, ctx);
+                              &p, inlen, (int)tt->tag, aclass, opt, ctx);
         q = p;
         if (!ret) {
             OPENSSL_PUT_ERROR(ASN1, ASN1_R_NESTED_ASN1_ERROR);
@@ -571,7 +571,7 @@ static int asn1_template_noexp_d2i(ASN1_VALUE **val,
     const unsigned char *p;
     if (!val)
         return 0;
-    flags = tt->flags;
+    flags = (int)tt->flags;
     aclass = flags & ASN1_TFLG_TAG_CLASS;
 
     p = *in;
@@ -582,7 +582,7 @@ static int asn1_template_noexp_d2i(ASN1_VALUE **val,
         char sk_eoc;
         /* First work out expected inner tag value */
         if (flags & ASN1_TFLG_IMPTAG) {
-            sktag = tt->tag;
+            sktag = (int)tt->tag;
             skaclass = aclass;
         } else {
             skaclass = V_ASN1_UNIVERSAL;
@@ -651,7 +651,7 @@ static int asn1_template_noexp_d2i(ASN1_VALUE **val,
         }
     } else if (flags & ASN1_TFLG_IMPTAG) {
         /* IMPLICIT tagging */
-        ret = asn1_item_ex_d2i(val, &p, len, ASN1_ITEM_ptr(tt->item), tt->tag,
+        ret = asn1_item_ex_d2i(val, &p, len, ASN1_ITEM_ptr(tt->item), (int)tt->tag,
                                aclass, opt, ctx, depth);
         if (!ret) {
             OPENSSL_PUT_ERROR(ASN1, ASN1_R_NESTED_ASN1_ERROR);
@@ -699,7 +699,7 @@ static int asn1_d2i_ex_primitive(ASN1_VALUE **pval,
         utype = tag;
         tag = -1;
     } else
-        utype = it->utype;
+        utype = (int)it->utype;
 
     if (utype == V_ASN1_ANY) {
         /* If type is ANY need to figure out type from tag */
@@ -798,7 +798,7 @@ static int asn1_d2i_ex_primitive(ASN1_VALUE **pval,
 
     /* We now have content length and type: translate into a structure */
     /* asn1_ex_c2i may reuse allocated buffer, and so sets free_cont to 0 */
-    if (!asn1_ex_c2i(pval, cont, len, utype, &free_cont, it))
+    if (!asn1_ex_c2i(pval, cont, (int)len, utype, &free_cont, it))
         goto err;
 
     *in = p;
@@ -1075,7 +1075,7 @@ static int collect_data(BUF_MEM *buf, const unsigned char **p, long plen)
 {
     int len;
     if (buf) {
-        len = buf->length;
+        len = (int)buf->length;
         if (!BUF_MEM_grow_clean(buf, len + plen)) {
             OPENSSL_PUT_ERROR(ASN1, ERR_R_MALLOC_FAILURE);
             return 0;
@@ -1133,7 +1133,7 @@ static int asn1_check_tlen(long *olen, int *otag, unsigned char *oclass,
             ctx->plen = plen;
             ctx->pclass = pclass;
             ctx->ptag = ptag;
-            ctx->hdrlen = p - q;
+            ctx->hdrlen = (int)(p - q);
             ctx->valid = 1;
             /*
              * If definite length, and no error, length + header can't exceed
