+++ Sources/CCryptoBoringSSL/crypto/asn1/tasn_enc.c
@@ -249,7 +249,7 @@ static int asn1_template_ex_i2d(ASN1_VALUE **pval, unsigned char **out,
 {
     int i, ret, flags, ttag, tclass;
     size_t j;
-    flags = tt->flags;
+    flags = (int)tt->flags;
     /*
      * Work out tag and class to use: tagging may come either from the
      * template or the arguments, not both because this would create
@@ -263,7 +263,7 @@ static int asn1_template_ex_i2d(ASN1_VALUE **pval, unsigned char **out,
             /* FIXME: error code here */
             return -1;
         /* Get tagging from template */
-        ttag = tt->tag;
+        ttag = (int)tt->tag;
         tclass = flags & ASN1_TFLG_TAG_CLASS;
     } else if (tag != -1) {
         /* No template tagging, get from arguments */
@@ -462,7 +462,7 @@ static int asn1_i2d_ex_primitive(ASN1_VALUE **pval, unsigned char **out,
     int utype;
     int usetag;
 
-    utype = it->utype;
+    utype = (int)it->utype;
 
     /*
      * Get length of content octets and maybe find out the underlying type.
